backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
SQLAlchemy
pandas
numpy
openpyxl
gunicorn
chardet
Werkzeug
psycopg2-binary
python-dotenv